VM daemon works in parallel with the pagedaemon threads, and, among other actions, swaps out kernel stacks of the processes.  On the other hand, currentl OOM logic which selects a process to kill in the critical condition, skips process with swapped-out thread.  Under some loads, this results in the big(gest) process being ignored by OOM.
Do not skip a process which has inhibited thread due to the swap-out,
in the OOM selection loop.  Note that killing such process requires
the thread stack page-in, but sometimes this is the only way to
recover.

Reported and tested by:	pho
Reviewed by:	alc
Sponsored by:	The FreeBSD Foundation
MFC after:	3 weeks


git-svn-id: https://svn.freebsd.org/base/head@290915 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f