Print more detailed info about the disk and partition chosen for booting. No behavioral changes, just cosmetics.
A partition number of zero is not a wildcard, it's the 'a' partition in
a BSD slice, so don't print it as "<auto>".  (Only slices are 1-based,
unit and partition numbers are 0-based and -1 is their wildcard marker.)

Also, after doing all the probing and choosing, print the final result as
"Booting from <disk spec>" where disk spec has all the wildcards resolved
and looks like familiar BSD slice-and-partition notation (disk0s3a, etc).


git-svn-id: https://svn.freebsd.org/base/head@291164 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f